refactor: replace deprecated comp
