+++ apps/web/src/views/Swap/V3Swap/containers/SwapCommitButton.tsx
@@ -1,15 +1,12 @@
-import { useTranslation } from '@pancakeswap/localization'
 import { TradeType } from '@pancakeswap/sdk'
 import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
-import { Currency, CurrencyAmount, Fraction, ONE, Percent } from '@pancakeswap/swap-sdk-core'
+import { Currency, Percent } from '@pancakeswap/swap-sdk-core'
 import { AutoColumn, Box, Button, Dots, Message, MessageText, Text, useModal } from '@pancakeswap/uikit'
-import { getUniversalRouterAddress } from '@pancakeswap/universal-router-sdk'
-import { useExpertMode } from '@pancakeswap/utils/user'
-import { confirmPriceImpactWithoutFee } from '@pancakeswap/widgets-internal'
-import { memo, useCallback, useEffect, useMemo, useState } from 'react'
-import { logGTMClickSwapEvent } from 'utils/customGTMEventTracking'
+import React, { memo, useCallback, useEffect, useMemo, useState } from 'react'
 
-import { getPermit2Address } from '@pancakeswap/permit2-sdk'
+import { useTranslation } from '@pancakeswap/localization'
+import { Permit2Signature, getUniversalRouterAddress } from '@pancakeswap/universal-router-sdk'
+import { confirmPriceImpactWithoutFee } from '@pancakeswap/widgets-internal'
 import { GreyCard } from 'components/Card'
 import { CommitButton } from 'components/CommitButton'
 import ConnectWalletButton from 'components/ConnectWalletButton'
@@ -23,22 +20,22 @@ import {
 } from 'config/constants/exchange'
 import { useCurrency } from 'hooks/Tokens'
 import { useIsTransactionUnsupported } from 'hooks/Trades'
-import { ApprovalState, useApproveCallback } from 'hooks/useApproveCallback'
-import usePermit2Allowance, { AllowanceState } from 'hooks/usePermit2Allowance'
-import useTransactionDeadline from 'hooks/useTransactionDeadline'
+import { usePermit } from 'hooks/usePermitStatus'
 import useWrapCallback, { WrapType } from 'hooks/useWrapCallback'
 import { Field } from 'state/swap/actions'
 import { useSwapState } from 'state/swap/hooks'
-import { useSwapActionHandlers } from 'state/swap/useSwapActionHandlers'
 import { useRoutingSettingChanged } from 'state/user/smartRouter'
 import { useCurrencyBalances } from 'state/wallet/hooks'
+import { logGTMClickSwapEvent } from 'utils/customGTMEventTracking'
 import { warningSeverity } from 'utils/exchange'
-import { isChainSupported } from 'utils/wagmi'
+import { isUserRejected } from 'utils/sentry'
 import { useAccount, useChainId } from 'wagmi'
-import { useParsedAmounts, useSlippageAdjustedAmounts, useSwapCallback, useSwapInputError } from '../hooks'
-import { useConfirmModalState } from '../hooks/useConfirmModalState'
+import { useParsedAmounts, useSlippageAdjustedAmounts, useSwapInputError } from '../hooks'
+import { useConfirmModalStateV2 } from '../hooks/useConfirmModalStateV2'
 import { TransactionRejectedError } from '../hooks/useSendSwapTransaction'
-import { useWallchainApi } from '../hooks/useWallchain'
+import { useSwapCallback } from '../hooks/useSwapCallback'
+import { useSwapConfig } from '../hooks/useSwapConfig'
+import { useSwapCurrency } from '../hooks/useSwapCurrency'
 import { computeTradePriceBreakdown } from '../utils/exchange'
 import { ConfirmSwapModal } from './ConfirmSwapModal'
 
@@ -50,53 +47,107 @@ interface SwapCommitButtonPropsType {
   tradeLoading?: boolean
 }
 
-export function useMaxAmountIn(trade: SmartRouterTrade<TradeType>, slippage: Percent, amountIn = trade?.inputAmount) {
-  if (!trade) return undefined
-  if (trade?.tradeType === TradeType.EXACT_INPUT) {
-    return amountIn
-  }
-
-  const slippageAdjustedAmountIn = new Fraction(ONE).add(slippage).multiply(amountIn.quotient).quotient
-  return CurrencyAmount.fromRawAmount(amountIn.currency, slippageAdjustedAmountIn)
+const useSettingModal = (onDismiss) => {
+  const [openSettingsModal] = useModal(
+    <SettingsModalWithCustomDismiss
+      customOnDismiss={onDismiss}
+      // customOnDismiss={() => setIndirectlyOpenConfirmModalState(true)}
+      mode={SettingsMode.SWAP_LIQUIDITY}
+    />,
+  )
+  return openSettingsModal
 }
 
-export const SwapCommitButton = memo(function SwapCommitButton({
-  trade,
-  tradeError,
-  tradeLoading,
-}: SwapCommitButtonPropsType) {
-  const { t } = useTranslation()
-  const { address: account } = useAccount()
-  const chainId = useChainId()
-  const [isExpertMode] = useExpertMode()
+const useSwapCurrencies = () => {
   const {
-    typedValue,
-    independentField,
     [Field.INPUT]: { currencyId: inputCurrencyId },
     [Field.OUTPUT]: { currencyId: outputCurrencyId },
   } = useSwapState()
   const inputCurrency = useCurrency(inputCurrencyId) as Currency
   const outputCurrency = useCurrency(outputCurrencyId) as Currency
-  const swapIsUnsupported = useIsTransactionUnsupported(inputCurrency, outputCurrency)
-  const { onUserInput } = useSwapActionHandlers()
+  return { inputCurrency, outputCurrency }
+}
+
+const WrapCommitButtonReplace: React.FC<React.PropsWithChildren> = ({ children }) => {
+  const { inputCurrency, outputCurrency } = useSwapCurrencies()
+  const { typedValue } = useSwapState()
   const {
     wrapType,
     execute: onWrap,
     inputError: wrapInputError,
   } = useWrapCallback(inputCurrency, outputCurrency, typedValue)
   const showWrap = wrapType !== WrapType.NOT_APPLICABLE
-  const [isRoutingSettingChange, resetRoutingSetting] = useRoutingSettingChanged()
+
+  if (!showWrap) return children
+
+  return (
+    <CommitButton width="100%" disabled={Boolean(wrapInputError)} onClick={onWrap}>
+      {wrapInputError ?? (wrapType === WrapType.WRAP ? 'Wrap' : wrapType === WrapType.UNWRAP ? 'Unwrap' : null)}
+    </CommitButton>
+  )
+}
+
+const ConnectButtonReplace = ({ children }) => {
+  const { address: account } = useAccount()
+
+  if (!account) {
+    return <ConnectWalletButton width="100%" />
+  }
+  return children
+}
+
+const UnsupportedSwapButtonReplace = ({ children }) => {
+  const { t } = useTranslation()
+  const { inputCurrency, outputCurrency } = useSwapCurrencies()
+  const swapIsUnsupported = useIsTransactionUnsupported(inputCurrency, outputCurrency)
+
+  if (swapIsUnsupported) {
+    return (
+      <Button width="100%" disabled>
+        {t('Unsupported Asset')}
+      </Button>
+    )
+  }
+  return children
+}
+
+export const SwapCommitButton: React.FC<SwapCommitButtonPropsType> = (props) => {
+  return (
+    <UnsupportedSwapButtonReplace>
+      <ConnectButtonReplace>
+        <WrapCommitButtonReplace>
+          <SwapCommitButtonInner {...props} />
+        </WrapCommitButtonReplace>
+      </ConnectButtonReplace>
+    </UnsupportedSwapButtonReplace>
+  )
+}
+
+const SwapCommitButtonInner = memo(function SwapCommitButtonInner({
+  trade,
+  tradeError,
+  tradeLoading,
+}: SwapCommitButtonPropsType) {
+  const { address: account } = useAccount()
+  const { t } = useTranslation()
+  const chainId = useChainId()
+  // form data
+  const { independentField } = useSwapState()
+  const [inputCurrency, outputCurrency] = useSwapCurrency()
+  const { isExpertMode, deadline } = useSwapConfig()
+
   const slippageAdjustedAmounts = useSlippageAdjustedAmounts(trade)
-  const deadline = useTransactionDeadline()
-  const [statusWallchain, approvalAddressForWallchain, wallchainMasterInput] = useWallchainApi(trade, deadline)
-  const [wallchainSecondaryStatus, setWallchainSecondaryStatus] = useState<'found' | 'not-found'>('not-found')
+  const amountToApprove = useMemo(
+    () => (inputCurrency?.isNative ? undefined : slippageAdjustedAmounts[Field.INPUT]),
+    [inputCurrency?.isNative, slippageAdjustedAmounts],
+  )
 
-  const routerAddress =
-    statusWallchain === 'found' || wallchainSecondaryStatus === 'found'
-      ? approvalAddressForWallchain
-      : getPermit2Address(chainId)
+  const tradePriceBreakdown = useMemo(() => computeTradePriceBreakdown(trade), [trade])
+  // warnings on slippage
+  const priceImpactSeverity = warningSeverity(
+    tradePriceBreakdown ? tradePriceBreakdown.priceImpactWithoutFee : undefined,
+  )
 
-  const amountToApprove = slippageAdjustedAmounts[Field.INPUT]
   const relevantTokenBalances = useCurrencyBalances(account ?? undefined, [
     inputCurrency ?? undefined,
     outputCurrency ?? undefined,
@@ -105,63 +156,31 @@ export const SwapCommitButton = memo(function SwapCommitButton({
     [Field.INPUT]: relevantTokenBalances[0],
     [Field.OUTPUT]: relevantTokenBalances[1],
   }
-
-  // check whether the user has approved the router on the input token
-  const { approvalState, currentAllowance, isPendingError } = useApproveCallback(amountToApprove, routerAddress)
-
-  const allowance = usePermit2Allowance(
-    routerAddress,
-    amountToApprove as any,
-    isChainSupported(chainId) ? getUniversalRouterAddress(chainId) : undefined,
-  )
-  // @ts-ignore
-  const tradePriceBreakdown = useMemo(() => !showWrap && computeTradePriceBreakdown(trade), [showWrap, trade])
+  const parsedAmounts = useParsedAmounts(trade, currencyBalances, false)
+  const parsedIndependentFieldAmount = parsedAmounts[independentField]
   const swapInputError = useSwapInputError(trade, currencyBalances)
-  const parsedAmounts = useParsedAmounts(trade, currencyBalances, showWrap)
-  const parsedIndepentFieldAmount = parsedAmounts[independentField]
 
-  // check if user has gone through approval process, used to show two step buttons, reset on token change
-  const [approvalSubmitted, setApprovalSubmitted] = useState<boolean>(false)
+  const [txHash, setTxHash] = useState<string | undefined>(undefined)
+  const [swapErrorMessage, setSwapErrorMessage] = useState<string | undefined>(undefined)
+  const [permit2Signature, setPermit2Signature] = useState<Permit2Signature | undefined>(undefined)
+  const [tradeToConfirm, setTradeToConfirm] = useState<SmartRouterTrade<TradeType> | undefined>(undefined)
+  const [indirectlyOpenConfirmModalState, setIndirectlyOpenConfirmModalState] = useState(false)
+  const statusWallchain: string = '@TODO'
 
-  const onWallchainDrop = useCallback(() => {
-    setApprovalSubmitted(false)
-  }, [setApprovalSubmitted])
-
+  const fn = useCallback(() => undefined, [])
   const {
     callback: swapCallback,
     error: swapCallbackError,
-    reason: revertReason,
+    reason: swapCallbackRevertReason,
   } = useSwapCallback({
     trade,
     deadline,
-    permitSignature: allowance.state === AllowanceState.ALLOWED ? allowance.permitSignature : undefined,
-    onWallchainDrop,
-    wallchainMasterInput,
-    statusWallchain,
+    permitSignature: permit2Signature,
+    onWallchainDrop: fn, // TODO
+    wallchainMasterInput: undefined, // TODO
+    statusWallchain: 'not-found', // TODO
   })
-
-  const [{ tradeToConfirm, swapErrorMessage, attemptingTxn, txHash }, setSwapState] = useState<{
-    tradeToConfirm: SmartRouterTrade<TradeType> | undefined
-    attemptingTxn: boolean
-    swapErrorMessage: string | undefined
-    txHash: string | undefined
-  }>({
-    tradeToConfirm: undefined,
-    attemptingTxn: false,
-    swapErrorMessage: undefined,
-    txHash: undefined,
-  })
-
-  // Handlers
-  const handleConfirmDismiss = useCallback(() => {
-    setSwapState({ tradeToConfirm, attemptingTxn, swapErrorMessage, txHash })
-    // if there was a tx hash, we want to clear the input
-    if (txHash) {
-      onUserInput(Field.INPUT, '')
-    }
-  }, [attemptingTxn, onUserInput, swapErrorMessage, tradeToConfirm, txHash, setSwapState])
-
-  const handleSwap = useCallback(async () => {
+  const onConfirmSwap = useCallback(async () => {
     if (
       tradePriceBreakdown &&
       !confirmPriceImpactWithoutFee(
@@ -173,215 +192,119 @@ export const SwapCommitButton = memo(function SwapCommitButton({
     ) {
       return undefined
     }
+
     if (!swapCallback) {
-      if (revertReason === 'insufficient allowance') {
-        setApprovalSubmitted(false)
-        setWallchainSecondaryStatus('found')
-        return undefined
+      if (swapCallbackRevertReason === 'insufficient allowance') {
+        // TODO
+        // setWallchainSecondaryStatus('found')
       }
       return undefined
     }
-    setSwapState({ attemptingTxn: true, tradeToConfirm, swapErrorMessage: undefined, txHash: undefined })
-    return swapCallback()
-      .then((res) => {
-        setWallchainSecondaryStatus('not-found')
-        setSwapState({ attemptingTxn: false, tradeToConfirm, swapErrorMessage: undefined, txHash: res.hash })
-        return res.hash
-      })
-      .catch((error) => {
-        console.error('swapCallback error#1001', error, JSON.stringify(error))
-        setWallchainSecondaryStatus('not-found')
+    try {
+      setTxHash(undefined)
+      const result = await swapCallback()
+      setTxHash(result.hash)
+      return result
+    } catch (error: any) {
+      console.error(error)
+      const userReject =
+        (typeof error !== 'string' && isUserRejected(error)) || error instanceof TransactionRejectedError
 
-        if (error instanceof TransactionRejectedError) {
-          setSwapState((s) => ({
-            ...s,
-            txHash: undefined,
-            attemptingTxn: false,
-          }))
-          // throw reject error to reset the flow
-          throw error
-        }
+      if (!userReject) setSwapErrorMessage(typeof error === 'string' ? error : error?.message)
 
-        setSwapState({
-          attemptingTxn: false,
-          tradeToConfirm,
-          swapErrorMessage:
-            (typeof error === 'string' ? error : error?.message) ??
-            t('Swap failed: %message%', { message: 'unknown reason #1001' }),
-          txHash: undefined,
-        })
-      })
-  }, [tradePriceBreakdown, t, swapCallback, tradeToConfirm, revertReason])
+      throw error
+    }
+  }, [swapCallback, swapCallbackRevertReason, t, tradePriceBreakdown])
+
+  const { execute: onStep, approvalState } = usePermit(
+    amountToApprove,
+    getUniversalRouterAddress(chainId),
+    onConfirmSwap,
+    permit2Signature,
+    setPermit2Signature,
+  )
+
+  const { confirmModalState, pendingModalSteps, resetConfirmModalState, startSwap } = useConfirmModalStateV2(
+    onStep,
+    amountToApprove,
+    approvalState,
+    permit2Signature,
+    setPermit2Signature,
+    getUniversalRouterAddress(chainId),
+  )
+  const reset = useCallback(() => {
+    resetConfirmModalState()
+    setTxHash(undefined)
+    setSwapErrorMessage(undefined)
+    setTradeToConfirm(undefined)
+    setPermit2Signature(undefined)
+  }, [resetConfirmModalState])
 
   const handleAcceptChanges = useCallback(() => {
-    setSwapState({ tradeToConfirm: trade, swapErrorMessage, txHash, attemptingTxn })
-  }, [attemptingTxn, swapErrorMessage, trade, txHash, setSwapState])
-  // End Handlers
+    setTradeToConfirm(trade)
+  }, [trade])
 
-  // warnings on slippage
-  const priceImpactSeverity = warningSeverity(
-    tradePriceBreakdown ? (tradePriceBreakdown.priceImpactWithoutFee as Percent) : undefined,
+  const noRoute = useMemo(() => !((trade?.routes?.length ?? 0) > 0) || tradeError, [trade?.routes?.length, tradeError])
+  const isValid = useMemo(() => !swapInputError && !tradeLoading, [swapInputError, tradeLoading])
+  const disabled = useMemo(
+    () =>
+      !isValid || (priceImpactSeverity > 3 && !isExpertMode) || !!swapCallbackError || statusWallchain === 'pending',
+    [isExpertMode, isValid, priceImpactSeverity, swapCallbackError],
   )
 
-  // show approve flow when: no error on inputs, not approved or pending, or approved in current session
-  // never show if price impact is above threshold in non expert mode
-  const showApproveFlow =
-    !swapInputError &&
-    (approvalState === ApprovalState.NOT_APPROVED ||
-      approvalState === ApprovalState.PENDING ||
-      allowance.state === AllowanceState.REQUIRED ||
-      allowance.state === AllowanceState.LOADING) &&
-    !(priceImpactSeverity > 3 && !isExpertMode)
-
-  // Modals
-  const [indirectlyOpenConfirmModalState, setIndirectlyOpenConfirmModalState] = useState(false)
-
-  const {
-    confirmModalState,
-    pendingModalSteps,
-    startSwapFlow,
-    onCancel: resetSwapFlow,
-  } = useConfirmModalState({
-    txHash,
-    chainId,
-    approval: approvalState,
-    approvalToken: trade?.inputAmount?.currency,
-    currentAllowance,
-    isExpertMode,
-    onConfirm: handleSwap,
-    allowance,
-  })
-
-  const [onPresentSettingsModal] = useModal(
-    <SettingsModalWithCustomDismiss
-      customOnDismiss={() => setIndirectlyOpenConfirmModalState(true)}
-      mode={SettingsMode.SWAP_LIQUIDITY}
-    />,
+  const userHasSpecifiedInputOutput = Boolean(
+    inputCurrency && outputCurrency && parsedIndependentFieldAmount?.greaterThan(BIG_INT_ZERO),
   )
 
-  const [onPresentConfirmModal] = useModal(
+  // modals
+  const onSettingModalDismiss = useCallback(() => {
+    setIndirectlyOpenConfirmModalState(true)
+  }, [])
+  const openSettingModal = useSettingModal(onSettingModalDismiss)
+  const [openConfirmSwapModal] = useModal(
     <ConfirmSwapModal
       trade={trade}
-      txHash={txHash}
-      approval={approvalState}
-      allowance={allowance}
-      attemptingTxn={attemptingTxn}
       originalTrade={tradeToConfirm}
-      showApproveFlow={showApproveFlow}
-      currencyBalances={currencyBalances}
-      swapErrorMessage={swapErrorMessage}
+      txHash={txHash}
       confirmModalState={confirmModalState}
       pendingModalSteps={pendingModalSteps}
-      isPendingError={isPendingError}
-      startSwapFlow={startSwapFlow}
+      swapErrorMessage={swapErrorMessage}
+      currencyBalances={currencyBalances}
       onAcceptChanges={handleAcceptChanges}
-      customOnDismiss={handleConfirmDismiss}
-      openSettingModal={onPresentSettingsModal}
+      onConfirm={startSwap}
+      openSettingModal={openSettingModal}
+      customOnDismiss={reset}
     />,
     true,
     true,
     'confirmSwapModal',
   )
-  // End Modals
 
-  const onSwapHandler = useCallback(() => {
-    setSwapState({
-      tradeToConfirm: trade,
-      attemptingTxn: false,
-      swapErrorMessage: undefined,
-      txHash: undefined,
-    })
-    resetSwapFlow()
+  const handleSwap = useCallback(() => {
+    setTradeToConfirm(trade)
+    resetConfirmModalState()
+
+    // if expert mode turn-on, will not show preview modal
+    // start swap directly
     if (isExpertMode) {
-      startSwapFlow()
+      startSwap()
     }
-    onPresentConfirmModal()
-    logGTMClickSwapEvent()
-  }, [trade, onPresentConfirmModal, isExpertMode, startSwapFlow, resetSwapFlow])
 
-  // useEffect
+    openConfirmSwapModal()
+    logGTMClickSwapEvent()
+  }, [isExpertMode, openConfirmSwapModal, resetConfirmModalState, startSwap, trade])
+
   useEffect(() => {
     if (indirectlyOpenConfirmModalState) {
       setIndirectlyOpenConfirmModalState(false)
-      setSwapState((state) => ({
-        ...state,
-        swapErrorMessage: undefined,
-      }))
-      onPresentConfirmModal()
+      setSwapErrorMessage(undefined)
+      openConfirmSwapModal()
     }
-  }, [indirectlyOpenConfirmModalState, onPresentConfirmModal, setSwapState])
-
-  // Reset approval flow if input currency changed
-  useEffect(() => {
-    setApprovalSubmitted(false)
-  }, [trade?.inputAmount?.currency])
-
-  // mark when a user has submitted an approval, reset onTokenSelection for input field
-  useEffect(() => {
-    if (approvalState === ApprovalState.PENDING) {
-      setApprovalSubmitted(true)
-    }
-  }, [approvalState, approvalSubmitted])
-
-  if (swapIsUnsupported) {
-    return (
-      <Button width="100%" disabled>
-        {t('Unsupported Asset')}
-      </Button>
-    )
-  }
-
-  if (!account) {
-    return <ConnectWalletButton width="100%" />
-  }
-
-  if (showWrap) {
-    return (
-      <CommitButton width="100%" disabled={Boolean(wrapInputError)} onClick={onWrap}>
-        {wrapInputError ?? (wrapType === WrapType.WRAP ? 'Wrap' : wrapType === WrapType.UNWRAP ? 'Unwrap' : null)}
-      </CommitButton>
-    )
-  }
-
-  const noRoute = !((trade?.routes?.length ?? 0) > 0) || tradeError
-
-  const userHasSpecifiedInputOutput = Boolean(
-    inputCurrency && outputCurrency && parsedIndepentFieldAmount?.greaterThan(BIG_INT_ZERO),
-  )
+  }, [indirectlyOpenConfirmModalState, openConfirmSwapModal])
 
   if (noRoute && userHasSpecifiedInputOutput && !tradeLoading) {
-    return (
-      <AutoColumn gap="12px">
-        <GreyCard style={{ textAlign: 'center', padding: '0.75rem' }}>
-          <Text color="textSubtle">{t('Insufficient liquidity for this trade.')}</Text>
-        </GreyCard>
-        {isRoutingSettingChange && (
-          <Message variant="warning" icon={<></>}>
-            <AutoColumn gap="8px">
-              <MessageText>{t('Unable to establish trading route due to customized routing.')}</MessageText>
-              <AutoRow gap="4px">
-                <RoutingSettingsButton
-                  buttonProps={{
-                    scale: 'xs',
-                    p: 0,
-                  }}
-                  showRedDot={false}
-                >
-                  {t('Check your settings')}
-                </RoutingSettingsButton>
-                <MessageText>{t('or')}</MessageText>
-                <Button variant="text" scale="xs" p="0" onClick={resetRoutingSetting}>
-                  {t('Reset to default')}
-                </Button>
-              </AutoRow>
-            </AutoColumn>
-          </Message>
-        )}
-      </AutoColumn>
-    )
+    return <ResetRoutesButton />
   }
-
-  const isValid = !swapInputError && !tradeLoading
 
   return (
     <Box mt="0.25rem">
@@ -389,10 +312,8 @@ export const SwapCommitButton = memo(function SwapCommitButton({
         id="swap-button"
         width="100%"
         variant={isValid && priceImpactSeverity > 2 && !swapCallbackError ? 'danger' : 'primary'}
-        disabled={
-          !isValid || (priceImpactSeverity > 3 && !isExpertMode) || !!swapCallbackError || statusWallchain === 'pending'
-        }
-        onClick={() => onSwapHandler()}
+        disabled={disabled}
+        onClick={handleSwap}
       >
         {swapInputError ||
           (tradeLoading && <Dots>{t('Searching For The Best Price')}</Dots>) ||
@@ -405,3 +326,37 @@ export const SwapCommitButton = memo(function SwapCommitButton({
     </Box>
   )
 })
+
+const ResetRoutesButton = () => {
+  const { t } = useTranslation()
+  const [isRoutingSettingChange, resetRoutingSetting] = useRoutingSettingChanged()
+  return (
+    <AutoColumn gap="12px">
+      <GreyCard style={{ textAlign: 'center', padding: '0.75rem' }}>
+        <Text color="textSubtle">{t('Insufficient liquidity for this trade.')}</Text>
+      </GreyCard>
+      {isRoutingSettingChange && (
+        <Message variant="warning" icon={<></>}>
+          <AutoColumn gap="8px">
+            <MessageText>{t('Unable to establish trading route due to customized routing.')}</MessageText>
+            <AutoRow gap="4px">
+              <RoutingSettingsButton
+                buttonProps={{
+                  scale: 'xs',
+                  p: 0,
+                }}
+                showRedDot={false}
+              >
+                {t('Check your settings')}
+              </RoutingSettingsButton>
+              <MessageText>{t('or')}</MessageText>
+              <Button variant="text" scale="xs" p="0" onClick={resetRoutingSetting}>
+                {t('Reset to default')}
+              </Button>
+            </AutoRow>
+          </AutoColumn>
+        </Message>
+      )}
+    </AutoColumn>
+  )
+}
