+++ apps/web/src/views/Swap/V3Swap/containers/index.ts
@@ -3,5 +3,4 @@ export * from './FormHeader'
 export * from './FormMain'
 export * from './PricingAndSlippage'
 export * from './SwapCommitButton'
-export * from './SwapCommitButtonV2'
 export * from './TradeDetails'
